Moved metrics endpoint to extension. Renamed to metrics
