+++ kie-server-parent/kie-server-remote/kie-server-rest/kie-server-rest-prometheus/src/main/resources/META-INF/services/org.kie.server.services.api.KieServerApplicationComponentsService
