+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-dmn/pom.xml
@@ -288,6 +288,11 @@
         </dependency>
         <dependency>
           <groupId>org.kie.server</groupId>
+          <artifactId>kie-server-rest-prometheus</artifactId>
+          <scope>test</scope>
+        </dependency>
+        <dependency>
+          <groupId>org.kie.server</groupId>
           <artifactId>kie-server-services-prometheus</artifactId>
           <scope>test</scope>
         </dependency>
