+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-dmn/src/test/java/org/kie/server/integrationtests/dmn/DMNMetricIntegrationTest.java
@@ -88,7 +88,7 @@ public class DMNMetricIntegrationTest
 
         Response response = null;
         try {
-            String uriString = TestConfig.getKieServerHttpUrl() + "/prometheus";
+            String uriString = TestConfig.getKieServerHttpUrl().replaceAll("/server", "") + "/metrics";
 
             WebTarget clientRequest = newRequest(uriString);
             response = clientRequest.request(MediaType.TEXT_PLAIN_TYPE).get();
