+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-dmn/src/test/resources/logback-test.xml
@@ -19,6 +19,7 @@
   <logger name="org.apache.http" level="WARN" />
   <logger name="org.apache.maven" level="WARN" />
   <logger name="org.kie.server.services.PrometheusMetricsDMNListener" level="DEBUG" />
+  <logger name="org.jboss.resteasy.core" level="DEBUG" />
 <!--
   <logger name="org.kie.server.services.dmn" level="DEBUG" />
   <logger name="org.kie.server.remote.rest.dmn" level="DEBUG" />
