+++ kie-server-parent/kie-server-wars/kie-server/src/main/assembly/assembly-ee8-container.xml
@@ -53,6 +53,7 @@
         <include>org.kie.server:kie-server-rest-optaplanner</include>
         <include>org.kie.server:kie-server-rest-dmn</include>
         <include>org.kie.server:kie-server-rest-swagger</include>
+        <include>org.kie.server:kie-server-rest-prometheus</include>
         <include>org.kie.server:kie-server-rest-case-mgmt</include>
         <include>org.kie.server:kie-server-router-client</include>
         <include>org.kie.server:kie-server-client</include>
